[fix] Result型からState型へ変更
