+++ android/app/src/main/java/com/wsr/index/IndexFragment.kt
@@ -10,7 +10,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
 import com.wsr.databinding.FragmentIndexBinding
-import com.wsr.utils.State
+import com.wsr.state.State
 import kotlinx.coroutines.launch
 
 class IndexFragment : Fragment() {
