+++ settings.gradle.kts
@@ -10,3 +10,4 @@
 rootProject.name = "Passon"
 include("core", "core:domain", "core:usecase", "core:repository")
 include("android", "android:app")
+include("utils")
